requirements.txt
torch
torchvision
torchaudio
transformers
opencv-python
pandas
numpy
python-Levenshtein